+++ sync-for-commerce/RELEASES.md
@@ -591,3 +591,13 @@ Based on:
 - [typescript v2.2.1] sync-for-commerce
 ### Releases
 - [NPM v2.2.1] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.2.1 - sync-for-commerce
+
+## 2023-09-29 00:07:12
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.0] sync-for-commerce
+### Releases
+- [NPM v2.3.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.3.0 - sync-for-commerce